requirements.txt
beautifulsoup4==4.13.3
Requests==2.32.3
python-dotenv==1.0.1
playwright==1.42.0
